EVERYTHING WORKS

+++ server.py
@@ -237,7 +237,14 @@ def login():
 def user_profile():
     # Retrieve the username from the query parameters
     username = request.args.get('username')
-    #username = request.get('usenname')
+    user_info = g.conn.execute(text("""
+        SELECT name, bio, image_url
+        FROM users
+        WHERE user_id = :username
+        """), {"username": username}
+    ).fetchone()
+
+
     # Check if the user is a personal profile
     personal_profile = g.conn.execute(
        text("SELECT * FROM personal_profile WHERE user_id = :username"), {"username": username}
@@ -248,7 +255,7 @@ def user_profile():
         posts = g.conn.execute(
             text("SELECT * FROM post WHERE user_id = :username"), {"username": username}
         ).fetchall()
-        return render_template('personal_profile.html', pp=personal_profile, posts=posts)
+        return render_template('personal_profile.html', u=user_info, pp=personal_profile, posts=posts)
 
     # Check if the user is a company profile
     company_profile = g.conn.execute(
@@ -264,8 +271,11 @@ def user_profile():
         job_listings = g.conn.execute(
             text("SELECT * FROM job_listing WHERE user_id = :username"), {"username": username}
         ).fetchall()
+        locations = g.conn.execute(
+            text("SELECT location FROM company_locations WHERE user_id = :username"), {"username": username}
+        ).fetchall()
 
-        return render_template('company_profile.html', user_name=company_profile['Name'], user_id=company_profile['User_id'], location=company_profile['Location'], bio=company_profile['Bio'], image=company_profile['Image_URL'], job_listings=job_listings)
+        return render_template('company_profile.html', u=user_info, cp=company_profile, posts=posts, jl=job_listings, cl=locations)
 
 
 @app.route('/groups', methods=['GET', 'POST'])
@@ -282,17 +292,17 @@ def groups():
 
         if action == 'join':
             # Insert into GROUP_MEMBERS table
-            g.conn.execute(text("INSERT INTO GROUP_MEMBERS (Group_id, User_id) VALUES (?, ?)"), (group_id, user_id))
+            g.conn.execute(text("INSERT INTO GROUP_MEMBERS (Group_id, User_id) VALUES (:group_id, :user_id)"), {"group_id": group_id, "user_id": user_id})
 
             # Update Number_of_members in GROUPS table
-            g.conn.execute(text("UPDATE GROUPS SET Number_of_members = Number_of_members + 1 WHERE Group_id = ?"), (group_id,))
+            g.conn.execute(text("UPDATE GROUPS SET Number_of_members = Number_of_members + 1 WHERE Group_id = :group_id"), {"group_id": group_id})
             
         elif action == 'leave':
             # Remove from GROUP_MEMBERS table
-            g.conn.execute(text("DELETE FROM GROUP_MEMBERS WHERE Group_id = ? AND User_id = ?"), (group_id, user_id))
+            g.conn.execute(text("DELETE FROM GROUP_MEMBERS WHERE Group_id = :group_id AND User_id = :user_id"), {"group_id": group_id, "user_id": user_id})
 
             # Update Number_of_members in GROUPS table
-            g.conn.execute(text("UPDATE GROUPS SET Number_of_members = Number_of_members - 1 WHERE Group_id = ?"), (group_id,))
+            g.conn.execute(text("UPDATE GROUPS SET Number_of_members = Number_of_members - 1 WHERE Group_id = :group_id"), {"group_id": group_id})
         
         return redirect(url_for('groups'))
 
@@ -305,26 +315,27 @@ def create_event():
         image_url = request.form['image_url']
         associated_date = request.form['associated_date']
         creation_date = date.today()
-
+        
         error = None
+        get_max_num = g.conn.execute(text("SELECT MAX(post_number) FROM post")).fetchone()
+        max_post_number = get_max_num[0] if get_max_num[0] is not None else 0
+        post_num = str(max_post_number + 1)
 
         try:
             g.conn.execute(
-                text('INSERT INTO POST (User_id, Creation_date, Image_URL, Text) VALUES (?, ?, ?, ?)'),
-                (user_id, creation_date, image_url, event_description)
+                text('INSERT INTO POST (User_id, Post_number, Creation_date, Image_URL, Text) VALUES (:user_id, :post_number, :creation_date, :image_url, :event_description)'),
+                {"user_id": user_id, "post_number": post_num, "creation_date": creation_date, "image_url": image_url, "event_description": event_description}
             )
             g.conn.execute(
-                text('INSERT INTO EVENT (User_id, Post_number, Associated_date) VALUES (?, (SELECT MAX(Post_number) FROM POST WHERE User_id = ?), ?)'),
-                (user_id, user_id, associated_date)
+                text('INSERT INTO EVENT (User_id, Post_number, Associated_date) VALUES (:user_id, (SELECT MAX(Post_number) FROM POST WHERE User_id = :user_id), :associated_date)'),
+                {"user_id": user_id, "associated_date": associated_date}
             )
-            g.conn.commit()
+            #g.conn.commit()
             flash('Event created successfully!', 'success')
         except Exception as e:
             error = str(e)
-            g.conn.rollback()
+            #g.conn.rollback()
             flash(f'An error occurred: {error}', 'error')
-
-        return redirect(url_for('index'))
       
     return render_template('event.html')
 
@@ -338,24 +349,25 @@ def announce():
         creation_date = date.today()
 
         error = None
+        get_max_num = g.conn.execute(text("SELECT MAX(post_number) FROM post")).fetchone()
+        max_post_number = get_max_num[0] if get_max_num[0] is not None else 0
+        post_num = str(max_post_number + 1)
 
         try:
             g.conn.execute(
-                text('INSERT INTO POST (User_id, Creation_date, Image_URL, Text) VALUES (?, ?, ?, ?)'),
-                (user_id, creation_date, image_url, announcement_text)
+                text('INSERT INTO POST (User_id, Post_number, Creation_date, Image_URL, Text) VALUES (:user_id, :post_number, :creation_date, :image_url, :announcement_text)'),
+                {"user_id": user_id, "post_number": post_num, "creation_date": creation_date, "image_url": image_url, "announcement_text": announcement_text}
             )
             g.conn.execute(
-                text('INSERT INTO ANNOUNCEMENT (User_id, Post_number) VALUES (?, (SELECT MAX(Post_number) FROM POST WHERE User_id = ?))'),
-                (user_id, user_id)
+                text('INSERT INTO ANNOUNCEMENT (User_id, Post_number) VALUES (:user_id, (SELECT MAX(Post_number) FROM POST WHERE User_id = :user_id))'),
+                {"user_id": user_id}
             )
-            g.conn.commit()
+            #g.conn.commit()
             flash('Announcement created successfully!', 'success')
         except Exception as e:
             error = str(e)
-            g.conn.rollback()
             flash(f'An error occurred: {error}', 'error')
 
-    
     return render_template('announce.html')
 
 
@@ -371,6 +383,9 @@ def feed():
     if request.method == 'POST':
         reaction = request.form['reaction']
         comment = request.form['comment']
+        post_owner_id = request.form['post_owner_id']  # Extract post_owner_id from form data
+        post_number = request.form['post_number']  # Extract post_number from form data
+
         # Insert reaction and comment into the database
         g.conn.execute(text("""
             INSERT INTO post_interaction (reaction, comment, post_owner_id, post_number, reacting_user_id)
@@ -378,17 +393,18 @@ def feed():
         """), {
             "reaction": reaction,
             "comment": comment,
-            "post_owner_id": request.form['post_owner_id'],
-            "post_number": request.form['post_id'],
+            "post_owner_id": post_owner_id,
+            "post_number": post_number,
             "reacting_user_id": user_id
         })
     
     posts_query = text("""  
         SELECT P.User_id AS Post_owner_id, P.Post_number, P.Creation_date AS Post_creation_date, P.Image_URL AS Post_image_url, P.Text AS Post_text, PI.Reaction, PI.Comment, PI.Reacting_user_id
         FROM Connect AS C
-        JOIN POST AS P ON (C.User_id2 = P.User_id AND C.Connection_status = 'Connected')
+        JOIN POST AS P ON ((C.User_id1 = P.User_id OR C.User_id2 = P.User_id) AND C.Connection_status = 'Connected')
         LEFT JOIN post_interaction AS PI ON (P.User_id = PI.Post_owner_id AND P.Post_number = PI.Post_number)
-        WHERE C.User_id1 = :user_id
+        WHERE :user_id IN (C.User_id1, C.User_id2)
+        ORDER BY P.Creation_date DESC
     """)
     post_out_raw = g.conn.execute(posts_query, {"user_id": user_id}).fetchall()
     post_out = [dict(row) for row in post_out_raw]
@@ -427,12 +443,12 @@ def for_you():
                 WHERE User_id = :user_id
             )
             OR U.User_id IN (
-                SELECT User_id
-                FROM COMPANY
-                WHERE Field = (
-                    SELECT Field
-                    FROM COMPANY
-                    WHERE User_id = :user_id
+                SELECT gm.User_id
+                FROM GROUP_MEMBERS gm
+                WHERE gm.Group_id IN (
+                    SELECT gm2.Group_id
+                    FROM GROUP_MEMBERS gm2
+                    WHERE gm2.User_id = :user_id
                 )
             )
         )
@@ -442,55 +458,69 @@ def for_you():
     return render_template('for_you.html', page_out=page_out)
 
 
-@app.route('/explore', methods=['GET','POST'])
+@app.route('/explore', methods=['GET', 'POST'])
 def search():
     if request.method == 'POST':
-        u_search = request.form['u_search']
-        l_search = request.form['l_search']
-        p_search = request.form['p_search']
-        if u_search:
-            u_results = g.conn.execute(text("""
-                SELECT user_id, name
-                FROM users AS U
-                WHERE name LIKE '%' || :keyword || '%' OR user_id LIKE '%' || :keyword || '%'
-                """), {"keyword": u_search}
-            ).fetchall()
-            return render_template('search.html', u_results=u_results)
-        elif l_search:
-            l_results = g.conn.execute(text("""
-                SELECT u.name, u.user_id
-                FROM users u
-                LEFT JOIN personal_profile pp ON u.user_id = pp.user_id
-                LEFT JOIN company_locations cl ON u.user_id = cl.user_id
-                WHERE u.name LIKE '%' || :keyword || '%' OR pp.location LIKE '%' || :keyword || '%' OR cl.location LIKE '%' || :keyword || '%'
-                """), {"keyword": l_search}
-            ).fetchall()
-            return render_template('search.html', l_results=l_results)
-        elif p_search:
-            p_results = g.conn.execute(text("""
-                SELECT user_id, name
-                FROM users AS U
-                WHERE name LIKE '%' || :u_search || '%' OR user_id LIKE '%' || :u_search || '%'
-                """), {"u_search": u_search}
-            ).fetchall()
-            return render_template('search.html', p_results=p_results)
+        # Get the search inputs from the form
+        u_search = request.form.get('u_search')
+        l_search = request.form.get('l_search')
+        p_search = request.form.get('p_search')
         
-    return render_template('search.html')
+        # Determine the search type based on the submitted form data
+        if u_search:
+            return redirect(url_for('search_results', search_type='users', query=u_search))
+        elif l_search:
+            return redirect(url_for('search_results', search_type='location', query=l_search))
+        elif p_search:
+            return redirect(url_for('search_results', search_type='position', query=p_search))
+        else:
+            return render_template('search.html', error_message='Please provide a search query.')
+
+    return render_template('search.html', error_message=None)
+
 
 @app.route('/search_results', methods=['GET'])
-def show_results():
-    results = request.args.get('results')
-    no_results = request.args.get('no_results')
+def search_results():
+    search_type = request.args.get('search_type')
+    query = request.args.get('query')
 
-    if results:
-        # Parse the JSON results and pass them to the template
-        results = json.loads(results)
-        return render_template('search_results.html', results=results)
-    elif no_results:
-        # If no results were found, render the template with a message
-        return render_template('search_results.html', no_results=True)
+    if search_type == 'users':
+        # Perform user search based on the query
+        u_results = g.conn.execute(text("""
+            SELECT user_id, name
+            FROM users AS U
+            WHERE name LIKE '%' || :keyword || '%' OR user_id LIKE '%' || :keyword || '%'
+            """), {"keyword": query}
+        ).fetchall()
+        return render_template('search_results.html', results=u_results)
+
+    elif search_type == 'location':
+        # Perform location search based on the query
+        l_results = g.conn.execute(text("""
+            SELECT u.name, u.user_id
+            FROM users u
+            LEFT JOIN personal_profile pp ON u.user_id = pp.user_id
+            LEFT JOIN company_locations cl ON u.user_id = cl.user_id
+            WHERE u.name LIKE '%' || :keyword || '%' OR pp.location LIKE '%' || :keyword || '%' OR cl.location LIKE '%' || :keyword || '%'
+            """), {"keyword": query}
+        ).fetchall()
+        return render_template('search_results.html', results=l_results)
+
+    elif search_type == 'position':
+        # Perform position search based on the query
+        p_results = g.conn.execute(text("""
+            SELECT DISTINCT COALESCE(PP.User_id, JL.User_id) AS User_id, U.Name
+            FROM USERS U
+            LEFT JOIN PERSONAL_PROFILE PP ON U.User_id = PP.User_id AND PP.Position LIKE '%' || :keyword || '%'
+            LEFT JOIN COMPANY C ON U.User_id = C.User_id
+            LEFT JOIN JOB_LISTING JL ON C.User_id = JL.User_id AND JL.Position LIKE '%' || :keyword || '%'
+            WHERE COALESCE(PP.Position, JL.Position) IS NOT NULL;
+            """), {"keyword": query}
+        ).fetchall()
+        return render_template('search_results.html', results=p_results)
+
     else:
-        # If no query parameters are present, redirect to the explore page
+        # If no search type is provided, redirect to the explore page
         return redirect(url_for('explore'))        
 
 
@@ -514,13 +544,18 @@ def conversation():
             )
         ORDER BY text_date;
     """), {'current_user': user_id, 'other_user': username}).fetchall()
+
     if request.method == 'POST':
         message = request.form['message']
-        g.conn.execute(text("""
-			INSERT INTO message (sender, receiver, text, text_date)
-            VALUES (:user_id, :username, :message, :date)
-			"""), {"user_id": user_id, "username": username, "message": message, "date": date.today()}
-            ).fetchall()
+        try:
+            g.conn.execute(text("""
+                INSERT INTO message (sender, receiver, text, text_date)
+                VALUES (:user_id, :username, :message, :date)
+            """), {"user_id": user_id, "username": username, "message": message, "date": date.today()})
+            flash('Message sent successfully!', 'success')
+        except Exception as e:
+            print("Error executing SQL query:", e)
+
 
     return render_template('conversation_with.html', messages=messages, other_user=username)
